Add effector types, separate player logic for server
