+++ src/client/player.rs
@@ -1,14 +1,12 @@
 use bevy::prelude::*;
-use bevy::sprite::Mesh2dHandle;
 use bevy::utils::HashMap;
 use blenvy::*;
 use client::*;
 use leafwing_input_manager::prelude::*;
 use lightyear::prelude::*;
 
-use crate::shared::input::{PlayerAction, ReplicateInputBundle};
+use crate::shared::input::{MovementSet, PlayerAction, ReplicateInputBundle};
 use crate::shared::player::{shared_handle_player_movement, PlayerId, PlayerMovement, SpaceShip};
-use crate::shared::MovementSet;
 
 use super::lobby::LobbyState;
 use super::MyClientId;
@@ -20,53 +18,13 @@ impl Plugin for PlayerPlugin {
         app.init_resource::<PlayerMap>()
             .add_systems(
                 Update,
-                (
-                    handle_player_spawn.run_if(resource_exists::<MyClientId>),
-                    convert_3d_to_2d_mesh,
-                    convert_std_to_color_material,
-                ),
+                (handle_player_spawn.run_if(resource_exists::<MyClientId>),),
             )
             .add_systems(
                 FixedUpdate,
                 handle_player_movement.in_set(MovementSet::Input),
             )
             .add_systems(OnEnter(LobbyState::None), despawn_input);
-    }
-}
-
-fn convert_3d_to_2d_mesh(mut commands: Commands, q_meshes: Query<(&Handle<Mesh>, &Name, Entity)>) {
-    for (mesh_handle, name, entity) in q_meshes.iter() {
-        commands
-            .entity(entity)
-            .remove::<Handle<Mesh>>()
-            .insert(Mesh2dHandle(mesh_handle.clone()));
-
-        info!("Converted {name:?} 3d mesh into 2d mesh.");
-    }
-}
-
-fn convert_std_to_color_material(
-    mut commands: Commands,
-    q_meshes: Query<(&Handle<StandardMaterial>, &Name, Entity)>,
-    std_materials: Res<Assets<StandardMaterial>>,
-    mut color_materials: ResMut<Assets<ColorMaterial>>,
-) {
-    for (std_material, name, entity) in q_meshes.iter() {
-        let Some(std_material) = std_materials.get(std_material) else {
-            continue;
-        };
-
-        let color_material = color_materials.add(ColorMaterial {
-            color: std_material.base_color,
-            texture: std_material.base_color_texture.clone(),
-        });
-
-        commands
-            .entity(entity)
-            .remove::<Handle<StandardMaterial>>()
-            .insert(color_material);
-
-        info!("Converted {name:?} standard material into color material.");
     }
 }
