+++ src/shared.rs
@@ -1,11 +1,13 @@
 //! This module contains the shared code between the client and the server.
-use bevy::{prelude::*, utils::Duration};
+use bevy::{prelude::*, sprite::Mesh2dHandle, utils::Duration};
 use blenvy::BlenvyPlugin;
+use input::MovementSet;
 use lightyear::prelude::*;
 
 pub const FIXED_TIMESTEP_HZ: f64 = 60.0;
 pub const SERVER_REPLICATION_INTERVAL: Duration = Duration::from_millis(100);
 
+pub mod effector;
 pub mod input;
 pub mod physics;
 pub mod player;
@@ -27,7 +29,50 @@ impl Plugin for SharedPlugin {
             crate::ui::UiPlugin,
             player::PlayerPlugin,
             physics::PhysicsPlugin,
-        ));
+            effector::EffectorPlugin,
+        ))
+        .add_systems(
+            Update,
+            (convert_3d_to_2d_mesh, convert_std_to_color_material),
+        );
+    }
+}
+
+/// Convert all 3d [`Handle<Mesh>`] to 2d [`Mesh2dHandle`].
+fn convert_3d_to_2d_mesh(mut commands: Commands, q_meshes: Query<(&Handle<Mesh>, &Name, Entity)>) {
+    for (mesh_handle, name, entity) in q_meshes.iter() {
+        commands
+            .entity(entity)
+            .remove::<Handle<Mesh>>()
+            .insert(Mesh2dHandle(mesh_handle.clone()));
+
+        info!("Converted {name:?} 3d mesh into 2d mesh.");
+    }
+}
+
+/// Convert all [`StandardMaterial`] to [`ColorMaterial`].
+fn convert_std_to_color_material(
+    mut commands: Commands,
+    q_meshes: Query<(&Handle<StandardMaterial>, &Name, Entity)>,
+    std_materials: Res<Assets<StandardMaterial>>,
+    mut color_materials: ResMut<Assets<ColorMaterial>>,
+) {
+    for (std_material, name, entity) in q_meshes.iter() {
+        let Some(std_material) = std_materials.get(std_material) else {
+            continue;
+        };
+
+        let color_material = color_materials.add(ColorMaterial {
+            color: Color::from(std_material.base_color.to_linear() + std_material.emissive),
+            texture: std_material.base_color_texture.clone(),
+        });
+
+        commands
+            .entity(entity)
+            .remove::<Handle<StandardMaterial>>()
+            .insert(color_material);
+
+        info!("Converted {name:?} standard material into color material.");
     }
 }
 
@@ -41,12 +86,4 @@ pub fn shared_config() -> SharedConfig {
         },
         mode: Mode::Separate,
     }
-}
-
-#[derive(SystemSet, Debug, Hash, PartialEq, Eq, Clone, Copy)]
-pub enum MovementSet {
-    // Input handling.
-    Input,
-    // Apply physics.
-    Physics,
 }
